StatusActivity now inflates the menu for StatusInfoNetApp due to changes in how TabActivity works in 3.0 and above.
